# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Perpustakaan",
    "description": "Perpustakaan API Collection",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3880/"
    }
  ],
  "paths": {
...
    "/koleksibuku/{id}": {
      "get": {
        "tags": [
          "Koleksibuku"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/menu/{id}": {
      "get": {
        "tags": [
          "Menu"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/webmenu": {
      "post": {
        "tags": [
          "Admin - Web Menu"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebmenuAdd"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Admin - Web Menu"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebmenuUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Web Menu"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebmenuDelete"
              }
            }
          }
        }
      }
    },
...
    "/webmenu/{id}": {
      "get": {
        "tags": [
          "Admin - Web Menu"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}